all the models are added separately, some issues left in Personal loan model
