+++ pipeline/pipeline_prod/requirements.txt
@@ -9,4 +9,5 @@ scikit-learn==1.3.1
 scipy==1.11.2
 matplotlib==3.8.0
 seaborn==0.12.2
+Flask
 google-auth
